Edit and delete of Image by owner only

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -93,13 +94,25 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
+      //Get image owner id
+        Integer imgOwnerId = image.getUser().getId();
+      //Get currently loggedin user id
+        User currentUser = (User) session.getAttribute("loggeduser");
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        String error = null;
+        if(currentUser.getId()==imgOwnerId) {
+        	model.addAttribute("editError", error);
+        	return "images/edit";
+        }
+        else {
+        	error = "Only the owner of the image can edit the image";
+        	model.addAttribute("editError", error);
+			return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -133,7 +146,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() + "/" + updatedImage.getTitle();
     }
 
 
@@ -141,10 +154,27 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
-    }
+	public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session,
+			RedirectAttributes redirectAttrs) {
+		Image image = imageService.getImage(imageId);
+	      //Get currently loggedin user id
+		User user = (User) session.getAttribute("loggeduser");
+		 //Get image owner id
+        Integer imgOwnerId = image.getUser().getId();
+		String error = null;
+		  if(user.getId()==imgOwnerId) {
+			imageService.deleteImage(imageId);
+			model.addAttribute("deleteError", error);
+			return "redirect:/images";
+		} else {
+			String imageTitle = image.getTitle();
+			error = "Only the owner of the image can delete the image";
+			String tags = convertTagsToString(image.getTags());
+			model.addAttribute("deleteError", error);
+			redirectAttrs.addAttribute("deleteError", error).addFlashAttribute("deleteError", error);
+			return "redirect:/images/" + imageId + '/' + imageTitle;
+		}
+	}
 
 
     //This method converts the image to Base64 format
@@ -187,5 +217,18 @@ public class ImageController {
         tagString.append(lastTag.getName());
 
         return tagString.toString();
+    }
+    
+    private Boolean imageOwner(Integer imageId, HttpSession session) {
+        //Get image details using image id
+        Image imgData = imageService.getImage(imageId);
+        //Get image owner id
+        Integer imgOwnerId = imgData.getUser().getId();
+        //Get currently loggedin user id
+        User currentUser = (User) session.getAttribute("loggeduser");
+        Integer currentUserId = currentUser.getId();
+
+        return currentUserId == imgOwnerId ? true : false;
+
     }
 }
